Awarding exponential scores.
Clearing mulitple lines at the same time is now awarded exponentially.
The increase in fitness value is now given by 2^n, where n is the number
of lines cleared during a move.

+++ code/include/phenotype.h
@@ -11,8 +11,8 @@ void free_phenotype (struct phenotype * phenotype);
 int compare_phenotypes (const void * p_1, const void * p_2);
 void write_phenotype (FILE * stream, struct phenotype * phenotype, struct options * opt);
 float board_score (struct board * new_board, struct board * old_board, struct phenotype* phenotype, struct t_last_placement * tlp, struct options * opt);
-int phenotype_fitness (struct phenotype* phenotype, struct options* opt);
-int average_phenotype_fitness (struct phenotype* phenotype, struct options* opt);
+unsigned long long int phenotype_fitness (struct phenotype* phenotype, struct options* opt);
+unsigned long long int average_phenotype_fitness (struct phenotype* phenotype, struct options* opt);
 int continue_board(struct board * board, struct phenotype* phenotype, int next_tetrominos[], struct options* opt);
 
 #endif
